Reorganise to improve OOP principles
